fix: Fix macOS build with Homebrew Clang and libc++
- Add CMake auto-detection of macOS SDK sysroot for Homebrew Clang
- Link against Homebrew's libc++ to resolve undefined symbols (__hash_memory)
- Add preprocessor guards for C++23 features missing in macOS libc++
  (std::views::stride, std::views::enumerate, std::chrono::clock_cast)
- Fix platform headers: use <util.h> instead of <pty.h>, _NSGetEnviron()
  instead of extern environ, add missing <sstream> include
- Force-include <algorithm> for llama.cpp b5460 on macOS
- Fix uninitialized variables (sigset_t, posix_spawn_file_actions_t)
- Add [[nodiscard]] attributes and explicit casts to suppress Clang warnings
- Rename signal handler members to satisfy clang-tidy naming convention
- Add NOLINT annotations for false-positive clang-analyzer diagnostics
- Fix %ld -> %lld format specifier for int64_t on macOS

+++ src/CoreVM/ir/IRBuilder.cpp
@@ -296,7 +296,7 @@ Value* IRBuilder::createPow(Value* lhs, Value* rhs, const std::string& name)
 
     if (auto* a = dynamic_cast<ConstantInt*>(lhs))
         if (auto* b = dynamic_cast<ConstantInt*>(rhs))
-            return get(static_cast<CoreNumber>(powl(a->get(), b->get())));
+            return get(static_cast<CoreNumber>(powl(static_cast<long double>(a->get()), static_cast<long double>(b->get()))));
 
     return insert<IPowInstr>(lhs, rhs, makeName(name));
 }
